factoring out some methods
