+++ libraries/reconstruction/movie_alignment_correlation_base.h
@@ -38,41 +38,13 @@
 #ifndef _PROG_MOVIE_ALIGNMENT_CORRELATION_BASE
 #define _PROG_MOVIE_ALIGNMENT_CORRELATION_BASE
 
+#include "data/alignment_result.h"
+#include "data/local_alignment_result.h"
 #include "core/xmipp_program.h"
 #include "core/metadata_extension.h"
 #include "core/xmipp_fftw.h"
-#include "data/point2D.h"
-#include "data/point3D.h"
-#include "data/rectangle.h"
 #include "core/optional.h"
-
-template<typename T>
-struct AlignmentResult {
-    size_t refFrame;
-    // these are shifts from the reference frame in X/Y dimension,
-    // i.e. if you want to compensate for the shift,
-    // you have to shift in opposite direction (negate these values)
-    std::vector<Point2D<T>> shifts;
-};
-
-template<typename T>
-struct FramePatchMeta {
-    // rectangle representing the patch
-    Rectangle<Point2D<T>> rec;
-    // logical id of the patch
-    size_t id_x;
-    size_t id_y;
-    size_t id_t;
-};
-
-template<typename T>
-struct LocalAlignmentResult {
-    const AlignmentResult<T> &globalHint;
-    // these are shifts (including global shift) of all patches in X/Y dimension,
-    // i.e. if you want to compensate for the shift,
-    // you have to shift in opposite direction (negate these values)
-    std::vector<std::pair<FramePatchMeta<T>, Point2D<T>>> shifts;
-};
+#include "eq_system_solver.h"
 
 template<typename T>
 class AProgMovieAlignmentCorrelation: public XmippProgram {
@@ -128,18 +100,6 @@ protected:
      */
     int findReferenceImage(size_t N, const Matrix1D<T>& shiftX,
             const Matrix1D<T>& shiftY);
-
-    /**
-     * Method computes absolute shifts from relative shifts
-     * @param bX relative shifts in X dim
-     * @param bY relative shifts in Y dim
-     * @param A system matrix to be used
-     * @param shiftX absolute shifts in X dim
-     * @param shiftY absolute shifts in Y dim
-     * @param verbose level
-     */
-    void solveEquationSystem(Matrix1D<T>& bX, Matrix1D<T>& bY, Matrix2D<T>& A,
-            Matrix1D<T>& shiftX, Matrix1D<T>& shiftY, int verbose);
 
     /**
      * Method to compute sum of shifts of some image in respect to a reference
@@ -389,6 +349,10 @@ protected:
     T outsideValue;
     /** if true, local alignment should be performed */
     bool processLocalShifts;
+    /** Solver iterations */
+    int solverIterations;
+    /** Metadata with shifts */
+    FileName fnOut;
 
 private:
     /** Filename of movie metadata */
@@ -399,10 +363,6 @@ private:
     T Ts;
     /** Max freq. */
     T maxFreq;
-    /** Solver iterations */
-    int solverIterations;
-    /** Metadata with shifts */
-    FileName fnOut;
     /** Do not calculate and use the input shifts */
     bool useInputShifts;
